#562: Make Parcels page controls sticky
